+++ rustfst/src/semirings/boolean_weight.rs
@@ -1,6 +1,6 @@
 use failure::Fallible;
 
-use crate::semirings::{CompleteSemiring, Semiring, SemiringProperties, StarSemiring};
+use crate::semirings::{CompleteSemiring, ReverseBack, Semiring, SemiringProperties, StarSemiring};
 use std::borrow::Borrow;
 /// Boolean semiring: (&, |, false, true).
 #[derive(Clone, Debug, PartialEq, PartialOrd, Default, Eq, Copy, Hash)]
@@ -54,6 +54,12 @@ impl Semiring for BooleanWeight {
             | SemiringProperties::COMMUTATIVE
             | SemiringProperties::IDEMPOTENT
             | SemiringProperties::PATH
+    }
+}
+
+impl ReverseBack<BooleanWeight> for BooleanWeight {
+    fn reverse_back(&self) -> Fallible<BooleanWeight> {
+        Ok(self.clone())
     }
 }