+++ rustfst/src/semirings/log_weight.rs
@@ -8,8 +8,8 @@ use ordered_float::OrderedFloat;
 
 use crate::parsers::bin_fst::utils_serialization::write_bin_f32;
 use crate::semirings::{
-    CompleteSemiring, DivideType, Semiring, SemiringProperties, SerializableSemiring, StarSemiring,
-    WeaklyDivisibleSemiring, WeightQuantize,
+    CompleteSemiring, DivideType, ReverseBack, Semiring, SemiringProperties, SerializableSemiring,
+    StarSemiring, WeaklyDivisibleSemiring, WeightQuantize,
 };
 use crate::KDELTA;
 use nom::number::complete::{float, le_f32};
@@ -94,6 +94,12 @@ impl Semiring for LogWeight {
         SemiringProperties::LEFT_SEMIRING
             | SemiringProperties::RIGHT_SEMIRING
             | SemiringProperties::COMMUTATIVE
+    }
+}
+
+impl ReverseBack<LogWeight> for LogWeight {
+    fn reverse_back(&self) -> Fallible<LogWeight> {
+        Ok(self.clone())
     }
 }