+++ rustfst/src/semirings/tropical_weight.rs
@@ -11,7 +11,7 @@ use ordered_float::OrderedFloat;
 use crate::parsers::bin_fst::utils_serialization::write_bin_f32;
 use crate::semirings::semiring::SerializableSemiring;
 use crate::semirings::{
-    CompleteSemiring, DivideType, Semiring, SemiringProperties, StarSemiring,
+    CompleteSemiring, DivideType, ReverseBack, Semiring, SemiringProperties, StarSemiring,
     WeaklyDivisibleSemiring, WeightQuantize,
 };
 use crate::KDELTA;
@@ -85,6 +85,12 @@ impl Semiring for TropicalWeight {
             | SemiringProperties::COMMUTATIVE
             | SemiringProperties::PATH
             | SemiringProperties::IDEMPOTENT
+    }
+}
+
+impl ReverseBack<TropicalWeight> for TropicalWeight {
+    fn reverse_back(&self) -> Fallible<TropicalWeight> {
+        Ok(self.clone())
     }
 }