+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `FstIteratorMut` as a trait bound of `MutableFst`
 - Add `take_final_weight` and `take_final_weight_unchecked` to MutableFst API.
 - Add `add_super_final_state` algorithm.
+- Add the `ReverseBack` trait that must be implemented for Semiring::ReverseWeight. Allows to avoid using transmute when we have `weight.reverse().reverse()` calls.
 
 ### Changed
 - `fst_convert` now consumes its input. Use `fst_convert_from_ref` to pass a borrow.